Make channel futures Send
Previous versions of the library couldn't be used with multithreaded
runtimes because the Futures haven't been marked as Send - even though
they technically were (had been synchronized by thread-safe mutexes).

Since however not all channels are thread-safe (LocalChannels are not)
it is not possible it is not possible to add the Send attribute to all
Future types. Therefore this change parameterizes the returned Futures
by the Mutex type and makes Futures only Sendable if a thread-safe
(Sync) mutex is used.

This implementation preferred to leak only the Mutex type into the
returned Future, and not the underlying channel type. This has the
advantage of a little bit less type leaks.