Yield in Mutex benchmarks
Async mutexes are intended to guard over longer time durations, which
lead to blocking a task. In order to simulate the tasks being blocked
for a longer time, the benchmark gets changed to let the task which
holds a Mutex to yield to the executor for a certain amount of time.