Wake Waker in Mutex outside of synchronous Mutex
Waking a Waker inside a synchronous Mutex can lead the next task to
get scheduled very quickly and immediately getting blocked on the
synchronous Mutex. In this case the thread executing this task might
need to get yield and scheduled again later. In order to avoid the
unnecessary scheduling operation we return the Waker and wake it
outside of the synchronous Mutex. In this case the next task will have
an increased chance to grab the synchronous Mutex.