Refactor the intrusive linked list
This is an attempt to reduce the amount of unsafe code and
raw pointers in intrusive list usage.

The main unsafe function is now adding a node to the intrusive linked
list, which requires the caller to assure they will remove the item
properly later on.

Many other functions have been converted to safe functions. In order for
the iterators to be safe, they had been converted to internal iterators.
External iterators did not provide proper guarantees on when they will
drain the list.

The change also moved from raw pointers to using
`Option<NonNull<ListNode<T>>>` for intrusive list